directiveToReactAdmon factored out to separate file
