+++ cli/extract/lib/nugget_catalog.js
@@ -14,6 +14,8 @@ const { visitParents } = require('unist-util-visit-parents')
 
 const { Nugget } = require('../../lib/nugget')
 
+const { directiveToReactAdmon } = require('./admonition')
+
 function modifyLinks ({ allNuggets, key }) {
   return (tree) => {
     visitParents(tree, 'image', (node, ancestors) => {
@@ -400,44 +402,6 @@ exports.NuggetCatalog = class NuggetCatalog {
       .use(remarkStringify, { resourceLink: true })
 
     return processor.processSync(markdown).toString()
-
-    function directiveToReactAdmon () {
-      return (tree) => {
-        visitParents(tree, 'containerDirective', (node, ancestors) => {
-          if (node.type === 'containerDirective') {
-            let title = ''
-
-            // pull the title from AST generated by :::note[title]
-            const children = node.children || (node.children = [])
-            if (children.length > 0 &&
-            children[0].type === 'paragraph' &&
-            children[0].data &&
-            children[0].data.directiveLabel &&
-            children[0].data.directiveLabel === true &&
-            children[0].children &&
-            children[0].children.length > 0 &&
-            children[0].children[0].type === 'text'
-            ) {
-              title = ` title="${children[0].children[0].value}"`
-              children[0].children.shift()
-            }
-
-            // include the Admonition in the children array
-            children.unshift({
-              type: 'html',
-              value: `<Admonition type="${node.name}"${title}>\n`
-            })
-
-            children.push({ type: 'html', value: '\n</Admonition>' })
-
-            // replace this containerDirective node with the children elements
-            const parent = ancestors[ancestors.length - 1]
-            const index = parent.children.indexOf(node)
-            parent.children.splice(index, 1, ...children)
-          }
-        })
-      }
-    }
   }
 
   // generate an MDX component named after the type
